% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piechartsonamap.R
\name{addpies}
\alias{addpies}
\title{Draw Pies on a Map}
\usage{
addpies(ggmap, piedata, circlesize = 0.1, legend = TRUE, legendloc = "TL")
}
\description{
http://stackoverflow.com/questions/10368180/plotting-pie-graphs-on-map-in-ggplot
}

