% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piechartsonamap.R
\name{makepie}
\alias{makepie}
\title{Createa PieChart from a Dataframe}
\usage{
makepie(df, trim = TRUE)
}
\arguments{
\item{df, }{dataframe. A five column dataframe where ID/Lat/Lon/Variable/Value}
}
\description{
Createa PieChart from a Dataframe
}

